feat: add GlobalExceptionHandler & rebase handlers
